Add support to VHAL properties
Based on libraries from Google (vhal_emulator), communicate with
SocketConn in AAOS side, being able to get and set VHAL properties.

Due to inconsistencies in AAOS open source builds, it is difficult
to execute the script. It is possible to get data from broker, but
since the interface on both emulators and Pixel devices is unstable,
crashes constantly come up during execution.

The code here can be considered a placeholder and used as a base
for future integration with AAOS, once the system car virtualisation
has a more defined and reliable interface.